Fixed complaints from cppcheck

+++ src/derep.cc
@@ -345,7 +345,7 @@ void dereplicate()
           fprint_id_noabundance(statsfile, sp->seqno_first);
           fprintf(statsfile, "\t%lu\t%u\t%u\t%u\n", 
                   db_getabundance(sp->seqno_first),
-                  sp->singletons, (unsigned int)0, (unsigned int)0);
+                  sp->singletons, 0U, 0U);
           progress_update(i);
         }
       progress_done();
